renderer running now again

+++ RayTracer/include/core/utility.h
@@ -11,7 +11,7 @@ inline glm::vec3 clamp(glm::vec3 v)
 	return glm::min(glm::vec3(1.f), glm::max(glm::vec3(0.f), v));
 }
 
-inline void crop(float min, float max, unsigned int x, unsigned int cropped[])
+inline void crop(float min, float max, size_t x, size_t cropped[])
 {
 	cropped[0] = int(round(clamp(min) * x));
 	cropped[1] = int(round(clamp(max) * x));
